Fixed previous checkout problem.
The Cause:
on line 228 in index.js, Sell Button was taking in the value of an input which always returns a string,  so if you bought two it would just add it to a string. It explains why it only happened to items that were added by sellers other than the store.

The Solution:
I used parseFloat() on line 229 in index.js, to change it to a float value before it was added to the INVENTORY.